requirements.txt
matplotlib==3.10.3
numpy==2.2.6
pyvisa==1.15.0
PyYAML==6.0.2
PyYAML==6.0.2
tqdm==4.67.1
